[WIP] Add generic tailer support, Tailer interface
The main complication here is in how tailers are stopped: docker and
socket both require interrupting ongoing read operations, and
coordinating shutdown of multiple goroutines.

This also required modifying the Start method to not take arguments (so
that it can be part of the interface signature) and renaming some
methods from WaitFlush to Stop.

TODO: file tailer