Convert pkg/logs/parser.Parser#Parse to return a struct, rather than a 5-tuple
This is a formaulaic change, although the added documentation is
important.

The performance implications of this change are likely to be minimal --
Go allocates space for a return tuple on the stack, and this is
allocating a Message on the stack.  And anyway, this is not a
performance-critical component of the logs agent.